Consolidate path normalization helpers into Core
Extract duplicated trimTrailingSeparators and isRootPath helpers from
lint workspace into @gml-modules/core. Both functions existed in two
separate files with slightly different implementations:

- src/lint/src/language/path-normalization.ts (simple version)
- src/lint/src/services/path-boundary.ts (comprehensive Windows/UNC version)

The comprehensive version with full Windows and UNC path support is now
the single source of truth in src/core/src/fs/path.ts. Both lint files
now import and use Core.trimTrailingSeparators, eliminating 55+ lines
of duplicated logic.
