+++ src/lint/src/services/path-boundary.ts
@@ -1,11 +1,9 @@
 import { realpathSync } from "node:fs";
 
+import { Core } from "@gml-modules/core";
+
 const WINDOWS_DRIVE_LETTER_PATTERN = /^[A-Za-z]:/;
-const WINDOWS_DRIVE_ROOT_PATTERN = /^[A-Za-z]:\\$/;
-const WINDOWS_DRIVE_ROOT_WITH_OPTIONAL_SEPARATOR_PATTERN = /^(?:[A-Za-z]:)\\?$/;
 const UNC_PREFIX_PATTERN = /^\\\\/;
-const UNC_SHARE_ROOT_PATTERN = /^\\\\[^\\]+\\[^\\]+$/;
-const UNC_SHARE_ROOT_WITH_TRAILING_SEPARATOR_PATTERN = /^\\\\[^\\]+\\[^\\]+\\$/;
 
 function isWindowsLikePath(value: string): boolean {
     return WINDOWS_DRIVE_LETTER_PATTERN.test(value) || UNC_PREFIX_PATTERN.test(value);
@@ -19,44 +17,6 @@ function normalizeSeparators(value: string): string {
     return value.replaceAll("\\", "/");
 }
 
-function isCanonicalRoot(value: string): boolean {
-    if (value === "/") {
-        return true;
-    }
-
-    if (WINDOWS_DRIVE_ROOT_PATTERN.test(value)) {
-        return true;
-    }
-
-    return UNC_SHARE_ROOT_PATTERN.test(value);
-}
-
-function normalizeWindowsRootShape(value: string): string {
-    if (WINDOWS_DRIVE_ROOT_WITH_OPTIONAL_SEPARATOR_PATTERN.test(value)) {
-        return `${value.slice(0, 2)}\\`;
-    }
-
-    if (UNC_SHARE_ROOT_WITH_TRAILING_SEPARATOR_PATTERN.test(value)) {
-        return value.slice(0, -1);
-    }
-
-    return value;
-}
-
-function trimTrailingSeparators(value: string): string {
-    const normalizedRoot = normalizeWindowsRootShape(value);
-    if (isCanonicalRoot(normalizedRoot)) {
-        return normalizedRoot;
-    }
-
-    let current = value;
-    while (current.endsWith("/") || current.endsWith("\\")) {
-        current = current.slice(0, -1);
-    }
-
-    return normalizeWindowsRootShape(current);
-}
-
 function canonicalizePathCase(value: string): string {
     if (isWindowsLikePath(value)) {
         return value.toLowerCase();
@@ -67,7 +27,7 @@ function canonicalizePathCase(value: string): string {
 
 function canonicalizeFromString(value: string): string {
     const withNormalizedSeparators = normalizeSeparators(value);
-    const trimmed = trimTrailingSeparators(withNormalizedSeparators);
+    const trimmed = Core.trimTrailingSeparators(withNormalizedSeparators);
     return canonicalizePathCase(trimmed);
 }
